feature-6/ rename TestUtil to ResourceTestHelper
